Make encode/decode utilities.
