+++ util/util_test.go
@@ -2,8 +2,50 @@ package util
 
 import (
 	"encoding/base64"
+	"io/ioutil"
+	"path"
 	"testing"
+
+	"github.com/stretchr/testify/require"
 )
+
+func TestEncodeToAndFromFile(t *testing.T) {
+	tempdir := t.TempDir()
+
+	type test struct {
+		One string `json:"one"`
+		Two int    `json:"two"`
+	}
+	data := test{
+		One: "one",
+		Two: 2,
+	}
+
+	{
+		pretty := path.Join(tempdir, "pretty.json")
+		err := EncodeToFile(pretty, data, true)
+		require.NoError(t, err)
+		require.FileExists(t, pretty)
+		var testDecode test
+		err = DecodeFromFile(pretty, &testDecode)
+		require.Equal(t, data, testDecode)
+
+		// Check the pretty printing
+		bytes, err := ioutil.ReadFile(pretty)
+		require.NoError(t, err)
+		require.Contains(t, string(bytes), "  \"one\": \"one\",\n")
+	}
+
+	{
+		small := path.Join(tempdir, "small.json")
+		err := EncodeToFile(small, data, false)
+		require.NoError(t, err)
+		require.FileExists(t, small)
+		var testDecode test
+		err = DecodeFromFile(small, &testDecode)
+		require.Equal(t, data, testDecode)
+	}
+}
 
 func TestPrintableUTF8OrEmpty(t *testing.T) {
 	encodeArg := func(str string) string {
